Disable container status modeling by default
Container status monitoring is disable by default because container down
events will only auto-clear if the same container is restarted. If the
container is left in a non-running state, or if is removed, its event
must be manually cleared. If auto-clearing is important you may want to
consider using Zenoss' normal process monitoring support to monitor the
process(es) running within the container instead of monitoring the
container.

Fixes ZEN-24043.