Restore Dynamic View for non-Docker components
The Docker ZenPack was a causing a problem with all other ZenPacks where
their components that should have a Dynamic View available didn't. It
turns out the reason for this was that the Docker ZenPack doesn't
include a custom device type, so the component Dynamic View for the
Docker ZenPack was being added to all devices in the system. Depending
on the order in which JavaScript snippets were loaded, this would either
result in Docker containers not having a Dynamic View, or no components
other than Docker containers having a Dynamic View.

This fix namespaces the Dynamic View component to the ZenPack. This
allows both Docker containers, and all other components to have a
Dynamic View without stomping on each other.

This is a general fix that will go into the ZenPackLib ZenPack for the
future too.

Fixes ZPS-703.